+++ src/components/Activity.js
@@ -2,7 +2,7 @@
 import Grid from '@material-ui/core/Grid';
 import { makeStyles } from '@material-ui/core/styles';
 import Typography from '@material-ui/core/Typography';
-import Top from './Top';
+import Top from './Top'
 import man from '../man.svg'
 import Avatar from '@material-ui/core/Avatar';
 import Box from '@material-ui/core/Box';
