+++ src/components/Myprofile.js
@@ -15,8 +15,11 @@ import { Link } from 'react-router-dom';
 import CheckRoundedIcon from '@material-ui/icons/CheckRounded';
 import Top from './Top';
 import { Add, School, Timelapse} from '@material-ui/icons';
-import SimpleMenu from './controls/SimpleMenu'
 import SimpleAccordion from './controls/SimpleAccordion';
+import ProjectForm from './Actions/ProjectForm';
+import ExperienceForm from './Actions/ExperienceForm'
+import CertificationForm from './Actions/CertificationForm'
+
 
 const useStyles = makeStyles((theme) => ({
 
@@ -79,26 +82,16 @@ const useStyles = makeStyles((theme) => ({
 function ProfileHeader() {
     const classes = useStyles();
 
-    const menuitems = [
-        {title:'Add Project', link:'/ProjectForm'},
-        {title:'Add Experience', link:'#'},
-        {title:'Add Certifications', link:'#'}
-    ]
+    const userdetails = {avatar: man, name: 'Kri'}
     return (
         <Grid container direction="row" justify="flex-start" alignItems="center">
-                        <Avatar alt="Remy Sharp" src={man} className={classes.large} />
-                        <Grid item xs={6} sm={3} alignItems="flex-start" justify="flex-start">
-                            <Typography color="textPrimary" variant="h6" >
-                                Name
-                            </Typography>
-                        </Grid>
-                        <Grid item xs={4}>
-                            <SimpleMenu items={menuitems}/>
-                        </Grid>
-                        <Grid item xs={2} sm={3}>
-                            <Button size="small" variant="outlined" component={Link} to="/AddProfileSection"><Add/></Button>
-                        </Grid>
-                    </Grid>
+            <Avatar alt="Remy Sharp" src={userdetails.avatar} className={classes.large} />
+            <Grid item xs={6} sm={3} alignItems="flex-start" justify="flex-start" style={{background: 'pink'}}>
+                <Typography color="textPrimary" variant="h6" align='left'>
+                    {userdetails.name}
+                </Typography>
+            </Grid>
+        </Grid>
     );
 }
 
@@ -176,12 +169,11 @@ function RecentActivities() {
 function DetailsAccordion (){
 
     const items = [
-        {title:'Posts',panel:'panel1', startIcon:<FolderOpenRoundedIcon fontSize="large"/>, entries: [{heading:"Abc", description:"Abcd", status:"Incomplete", statusIcon:<Timelapse/>}, {heading:"Abc", description:"Abcd", status:"Completed", statusIcon:<CheckRoundedIcon/>}]},
-        {title:'Experiences',panel:'panel2', startIcon:<WorkOutline fontSize="large"/>, entries: [{heading:"Abc", description:"Abcd", status:"Compl", statusIcon:<CheckRoundedIcon/>}]},
-        {title:'Education',panel:'panel3', startIcon:<School fontSize="large"/>, entries: [{heading:"Abc", description:"Abcd", status:"Compl", statusIcon:<CheckRoundedIcon/>}]},
-        {title:'Certifications',panel:'panel4', startIcon:<School fontSize="large"/>, entries: [{heading:"Abc", description:"Abcd", status:"Compl", statusIcon:<CheckRoundedIcon/>}]},
-        {title:'Projects',panel:'panel5', startIcon:<School fontSize="large"/>, entries: [{heading:"Abc", description:"Abcd", status:"Compl", statusIcon:<CheckRoundedIcon/>}]},
-        
+        {form: <ProjectForm/>,title:'Posts',panel:'panel1', startIcon:<FolderOpenRoundedIcon fontSize="large"/>, entries: [{heading:"Abc", description:"Abcd", status:"Incomplete", statusIcon:<Timelapse/>}, {heading:"Abc", description:"Abcd", status:"Completed", statusIcon:<CheckRoundedIcon/>}]},
+        {form: <ExperienceForm/>,title:'Experiences',panel:'panel2', startIcon:<WorkOutline fontSize="large"/>, entries: [{heading:"Abc", description:"Abcd", status:"Compl", statusIcon:<CheckRoundedIcon/>}]},
+        {form: <ProjectForm/>,title:'Education',panel:'panel3', startIcon:<School fontSize="large"/>, entries: [{heading:"Abc", description:"Abcd", status:"Compl", statusIcon:<CheckRoundedIcon/>}]},
+        {form: <CertificationForm/>,title:'Certifications',panel:'panel4', startIcon:<School fontSize="large"/>, entries: [{heading:"Abc", description:"Abcd", status:"Compl", statusIcon:<CheckRoundedIcon/>}]},
+        {form: <ProjectForm/>,title:'Projects',panel:'panel5', startIcon:<School fontSize="large"/>, entries: [{heading:"Abc", description:"Abcd", status:"Compl", statusIcon:<CheckRoundedIcon/>}]},    
     ]
 
     const classes = useStyles();
